Handle static_keys for inter-section jump
Applying static_keys to an entry where the target jump is to a function
from another section will result in an incorrect target address being
calculated. To correctly calculate the relative address for a jump
instruction, the difference from the absolute address between the jump
instruction address and the target address must be taken.