Detecting ambiguous symbols in .data..once section
The .data..once section lacks a related relocation section, so it's not
possible to search for ambiguous symbols in this section. Instead of
relocations, symbols that refer to this section must be found by
directly traversing the .symtab section by checking st_shndx value.
Additionally, the symbol names are formatted as <NAME>.__already_done.X,
which doesn't match the current search pattern.