Handle kCFI
When the modified function makes an indirect function call (calls a
function through a pointer), and kernel control flow integrity (kCFI) is
enabled, the kernel crashes.

The investigation reveals that a crash is caused by a mismatched
calculated magic tag that is checked before every indirect call, which
is a kCFI mitigation, in modified function. Additionally the cfi prefix
function needs to be also included in the livepatch module for each
modified function.

To address this issue, every ud2 instruction that precedes an indirect
function call should be replaced with nop in generated livepatch module.